requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
dj-rest-auth
gunicorn